+++ browser/src/neovim/SharedNeovimInstance.ts
@@ -78,10 +78,6 @@ export class MenuBinding extends Binding implements IMenuBinding {
 
     public async setItems(items: string[], activeId?: string): Promise<void> {
         this._currentOptions = items
-        console.dir(this._currentOptions)
-//         this._currentOptions = options
-
-//         await this._initPromise
 
         const currentWinId = await this.neovimInstance.request("nvim_get_current_win", [])
         const currentBufId = await this.neovimInstance.eval("bufnr('%')")
@@ -104,20 +100,10 @@ export class MenuBinding extends Binding implements IMenuBinding {
     }
 }
 
-// TODO: We should not be making multiple instances of this class for each menu UI
-// Need to come with a paradigm to reuse them across instances (attach/detach)
-
 class SharedNeovimInstance implements SharedNeovimInstance {
 
     private _initPromise: Promise<void>
     private _neovimInstance: NeovimInstance
-
-    // private _currentOptions: Array<INeovimMenuOption<T>> = []
-    // private _cursorPositionChangedEvent: Event<INeovimMenuOption<T>> = new Event<INeovimMenuOption<T>>()
-
-    // public get onCursorPositionChanged(): IEvent<INeovimMenuOption<T>> {
-    //     return this._cursorPositionChangedEvent
-    // }
 
     public bindToMenu(): IMenuBinding {
         return new MenuBinding(this._neovimInstance)
@@ -125,16 +111,6 @@ class SharedNeovimInstance implements SharedNeovimInstance {
 
     constructor() {
         this._neovimInstance = new NeovimInstance(5, 5)
-
-        // this._neovimInstance.on("event", (eventName: string, evt: any) => {
-        //     const line = evt.line - 1
-
-        //     if (eventName === "CursorMoved") {
-        //         if (line < this._currentOptions.length) {
-        //             this._cursorPositionChangedEvent.dispatch(this._currentOptions[line])
-        //         }
-        //     }
-        // })
 
         this._neovimInstance.onOniCommand.subscribe((command: string) => {
             commandManager.executeCommand(command)
@@ -153,37 +129,18 @@ class SharedNeovimInstance implements SharedNeovimInstance {
         await this._initPromise
         console.log("SharedNeovimInstance started!")
     }
-
-    // public async input(input: string): Promise<void> {
-    //     await this._neovimInstance.input(input)
-    // }
-
-//     public async setOptions(options: Array<INeovimMenuOption<T>>): Promise<void> {
-
-//         this._currentOptions = options
-
-//         await this._initPromise
-
-//         const currentBufId = await this._neovimInstance.eval("bufnr('%')")
-//         const bufferLength = await this._neovimInstance.eval<number>("line('$')")
-
-//         const elems = []
-
-//         for (let i = 0; i < this._currentOptions.length; i++) {
-//             elems.push(i.toString())
-//         }
-
-//         await this._neovimInstance.request("nvim_buf_set_lines", [currentBufId, 0, bufferLength - 1, false, elems])
-//     }
 }
 
 let _sharedInstance: SharedNeovimInstance = null
-export const getInstance = (): SharedNeovimInstance => {
-
-    if (!_sharedInstance) {
-        _sharedInstance = new SharedNeovimInstance()
-        _sharedInstance.start()
+export const activate = async (): Promise<void> => {
+    if (_sharedInstance) {
+        return
     }
 
+    _sharedInstance = new SharedNeovimInstance()
+    await _sharedInstance.start()
+}
+
+export const getInstance = (): SharedNeovimInstance => {
     return _sharedInstance
 }
